--- specification/watcher/_types/EmailAction.ts
@@ -1,58 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { Id } from '@_types/common'
-import { DateString } from '@_types/Time'
-
-export enum DataAttachmentFormat {
-  json = 0,
-  yaml = 1
-}
-
-export class EmailBody {
-  html: string
-  text: string
-}
-
-export enum EmailPriority {
-  lowest = 0,
-  low = 1,
-  normal = 2,
-  high = 3,
-  highest = 4
-}
-
-export class EmailActionResult {
-  account?: string
-  message: EmailResult
-  reason?: string // TODO: not sure if this needed
-}
-
-export class EmailResult {
-  bcc?: string[]
-  body?: EmailBody
-  cc?: string[]
-  from?: string
-  id: Id
-  priority?: EmailPriority
-  reply_to?: string[]
-  sent_date: DateString
-  subject: string
-  to: string[]
-}
